blackletter/blackletter-upstream/requirements.txt
# Core web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.0

# Background tasks
celery==5.3.4
redis==5.0.1

# Object storage
boto3==1.34.0
botocore==1.34.0

# Document processing
PyMuPDF==1.23.14
python-docx==1.1.0
pytesseract==0.3.10
Pillow==10.1.0

# PDF generation
weasyprint==60.2
jinja2==3.1.2

# File type detection
python-magic==0.4.27

# LLM integration
openai==1.6.0

# Configuration
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Utilities
PyYAML==6.0.1
httpx==0.25.2

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.0.0
